Refactor: Organize App.tsx into Smaller Components
