+++ site/package.json
@@ -18,7 +18,8 @@
     "moment-timezone": "^0.5.44",
     "react": "^17.0.2",
     "react-datetime": "^3.2.0",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^6.22.3"
   },
   "devDependencies": {
     "@babel/core": "^7.17.5",
@@ -32,6 +33,7 @@
     "html-webpack-plugin": "^5.5.0",
     "style-loader": "^3.3.1",
     "ts-loader": "^9.5.1",
+    "typescript": "^5.4.2",
     "webpack": "^5.76.0",
     "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^4.15.1"
